CIS 1.20: Only include failed regions in evidence
Otherwise, we are creating a lot of new fields which lead to
elasticsearch errors:

```json
{
  "log.level": "warn",
  "@timestamp": "2023-04-30T09:19:30.121Z",
  "log.logger": "elasticsearch",
  "log.origin": {
    "file.name": "elasticsearch/client.go",
    "file.line": 429
  },
  "message": "Cannot index event publisher.Event{...<event data>...} (status=400): {\"type\":\"document_parsing_exception\",\"reason\":\"[1:6485] failed to parse: Limit of total fields [1000] has been exceeded while adding new fields [146]\",\"caused_by\":{\"type\":\"illegal_argument_exception\",\"reason\":\"Limit of total fields [1000] has been exceeded while adding new fields [146]\"}}, dropping event!",
  "service.name": "cloudbeat",
  "ecs.version": "1.6.0"
}
```

This is because every region in the map gets its own fields.

This is the simplest solution for now, because it's not possible to
flatten the map into an array of all analyzers because we need to know
which analyzers are missing for each region.