Fix array_contains to match string exactly (#228)
Previous behaviour: `contains(array[_], key)` checks if any string which
is an element of the array contains the provided `key` as a substring.

Correct behaviour: `array[_] == key` checks if any string which is an
element of the array matches exactly the provided `key`.

I think the name of the function `array_contains` hints towards the
second behaviour and that's why the function was misused in #219. I
found the bug while implementing cloudbeat functional tests for CIS AWS
1.16 and saw that `AmazonEC2FullAccess` (a default AWS policy) failed
the rule.

I've checked the usage of the function in other places and I don't think
my change can break previous expected behaviour. However, I don't think
there was any other latent bug in other use cases because this function
was used to search for enum values and not arbitrary strings.